+++ src/components/Tooltip/index.js
@@ -1,32 +1,163 @@
 // @flow
-import React, { type Node } from 'react';
-import TooltipWrapper from './TooltipWrapper';
+import React, { PureComponent } from 'react';
+import styled from 'styled-components';
 
-type TooltipVariant = 'up' | 'down' | 'left' | 'right';
+import Arrow from './Arrow';
+import TooltipWrapper from './Tooltip';
+import Bubble from './Bubble';
+import type { FadeEasing } from './Tooltip';
+
+export type Placement = 'left' | 'right' | 'top' | 'bottom';
+export type TooltipBehavior = 'hover' | 'click' | 'ref';
 
 type Props = {
-  placement: TooltipVariant,
-  dataTip: string,
-  children: Node,
-  /** A className can be passed down for further styling or extending with CSS-in-JS */
-  className?: string,
-  /** Defaults to theme secondary color if no value is passed */
+  behavior?: TooltipBehavior,
+  durationOnClick?: number,
+  arrowWidth?: number,
+  background?: string,
+  border?: string,
+  children?: Node,
   color?: string,
-  style?: Object,
+  content: Node,
+  className?: string,
+  style?: object,
+  fadeDuration?: number,
+  fadeEasing?: FadeEasing,
+  fixed?: boolean,
+  fontFamily?: string,
+  fontSize?: string,
+  offset?: number,
+  padding?: number,
+  placement?: Placement,
+  radius?: number,
+  zIndex?: number,
 };
 
-const Tooltip = ({ placement, dataTip, children, color, className, style }: Props) => (
-  <TooltipWrapper color={color} className={className} style={style}>
-    <span tooltip={dataTip} flow={placement}>
-      {children}
-    </span>
-  </TooltipWrapper>
-);
+type State = {
+  isOpen: boolean,
+};
+
+const Container = styled.div`
+  position: relative;
+  display: inline-block;
+`;
+
+class Tooltip extends PureComponent<Props, State> {
+  state = {
+    isOpen: false,
+  };
+
+  timeDelay = (msDuration: number) =>
+    new Promise(resolve => setTimeout(() => resolve(true), msDuration));
+
+  showTooltip = () => {
+    this.setState({ isOpen: true });
+  };
+
+  hideTooltip = () => {
+    this.setState({ isOpen: false });
+  };
+
+  flashTooltip = async () => {
+    const { durationOnClick } = this.props;
+    const { isOpen } = this.state;
+
+    if (isOpen) {
+      return;
+    }
+
+    this.showTooltip();
+    await this.timeDelay(durationOnClick);
+    this.hideTooltip();
+  };
+
+  render() {
+    const { isOpen } = this.state;
+    const {
+      behavior,
+      arrowWidth,
+      background,
+      border,
+      children,
+      color,
+      content,
+      style,
+      className,
+      fadeDuration,
+      fadeEasing,
+      fixed,
+      fontFamily,
+      fontSize,
+      offset,
+      padding,
+      placement,
+      radius,
+      zIndex,
+    } = this.props;
+
+    const tooltipElement = (
+      <TooltipWrapper
+        isOpen={!children || fixed ? true : isOpen}
+        placement={placement}
+        offset={offset + arrowWidth}
+        zIndex={zIndex}
+        fadeEasing={fadeEasing}
+        fadeDuration={fadeDuration}
+      >
+        <Bubble
+          background={background}
+          border={border}
+          color={color}
+          radius={radius}
+          fontFamily={fontFamily}
+          fontSize={fontSize}
+          padding={padding}
+        >
+          <Arrow width={arrowWidth} background={background} border={border} placement={placement} />
+          {content}
+        </Bubble>
+      </TooltipWrapper>
+    );
+
+    return children ? (
+      <Container
+        style={style}
+        className={className}
+        onClick={behavior === 'click' ? this.flashTooltip : undefined}
+        onMouseEnter={behavior === 'hover' && !fixed ? this.showTooltip : undefined}
+        onMouseLeave={behavior === 'hover' && !fixed ? this.hideTooltip : undefined}
+      >
+        {children}
+        {tooltipElement}
+      </Container>
+    ) : (
+      <Container className={className} style={style}>
+        {tooltipElement}
+      </Container>
+    );
+  }
+}
 
 Tooltip.defaultProps = {
-  className: '',
-  color: null,
+  behavior: 'hover',
+  durationOnClick: 1000,
+  arrowWidth: 8,
+  background: '',
+  border: '',
+  children: null,
+  color: '',
+  fadeDuration: 150,
+  fadeEasing: 'fade-in-out',
+  fixed: false,
+  fontFamily: 'inherit',
+  fontSize: null,
+  offset: 0,
+  padding: 0.7,
+  placement: 'top',
+  radius: 0,
+  zIndex: 1,
   style: {},
+  className: null,
 };
 
 export default Tooltip;