+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 
 =======
 
+# [0.16.0] - 21-05-2019
+
+### Added
+
+- Step Wizard
+
+### Change
+
+- Modify Tooltip behavior
+
+# [0.15.2] - 16-05-2019
+
+### Fixes
+
+- Make endDate prop optional in DatePicker component
+
 # [0.15.1] - 15-05-2019
 
 ### Change